pyproject.toml
[project]
name = "food-diary"
version = "0.1.0"
description = "A minimal Python web app using Starlette and Pug templates."
readme = "README.md" # You might want to create this file
requires-python = ">=3.8"
dependencies = [
  "starlette",
  "uvicorn[standard]", # [standard] includes common extras like websockets and http tools
  "pypugjs",
  "authlib",           # OAuth client library
  "httpx",             # Required by authlib for HTTP requests
  "itsdangerous",      # Session management and security
  "python-dotenv",     # Environment variable management
  "mangum",            # ASGI adapter for AWS Lambda
  "psycopg2-binary",   # PostgreSQL adapter
]

[project.urls]
Homepage = "https://example.com" # Replace with your project's homepage if any
# Issues = "https://github.com/user/project/issues" # Replace if you have an issue tracker

# Optional: If you want to make your app runnable with `python -m <your_package_name>`
# or define console scripts, you can add:
# [project.scripts]
# my-app = "main:app" # Assuming your Starlette app instance is `app` in `main.py`
# This would require structuring your project as a package.
# For now, running with `uvicorn main:app` is simpler.

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["food_diary"]

[project.optional-dependencies]
test = [
  "pytest",     # Keep for unit tests with test_main.py
  "httpx",      # TestClient uses httpx
  "playwright",
  "nose2",      # Add nose2 for E2E tests
]
cdk = ["aws-cdk-lib>=2.0.0", "constructs>=10.0.0", "boto3"]

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 100

[tool.ruff.lint]
extend-select = ["E", "F", "I"]
unfixable = ["F841"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
known-first-party = ["food_diary"]
